Studies service cleanup
* Split out error into a separate class file
* Use the cached JDBI classes
